Criado serializers.py pra centralizar os serializers e ajustado o paylod pra retornar novo access_token e refresh_token.
